Add the fixed package
