+++ src/test/java/za/co/tt/controllerTest/AddressControllerTest.java
@@ -2,7 +2,7 @@
  * Author: Yanga Jilaji
  * Student number: 222582731
  * */
-package controllerTest;
+package za.co.tt.controllerTest;
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
@@ -14,9 +14,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import za.co.tt.Main;
-import za.co.tt.domain.User;
-import za.co.tt.domain.UserRole;
-import za.co.tt.factory.UserFactory;
+import za.co.tt.domain.Address;
+import za.co.tt.factory.AddressFactory;
 
 import java.time.LocalDate;
 
@@ -24,47 +23,57 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @TestMethodOrder(MethodOrderer.MethodName.class)
 @SpringBootTest(classes = Main.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class UserControllerTest {
+public class AddressControllerTest {
 
-    private static User user = UserFactory.createUser(
-            1001L, "Yanga", "Jilaji", "yjilaji@gmail.com", "pass123", "0682941085",
-            LocalDate.now(), true, null, null, UserRole.CUSTOMER);
+    private static Address address = AddressFactory.createAddress(
+            110L, "St Marks Road", "Cape Town", "Western Cape", 8001,
+            "South Africa", true, null, LocalDate.now(), LocalDate.now());
 
     @Autowired
     private TestRestTemplate restTemplate;
 
-    private final String baseUrl = "http://localhost:8080/user";
+    private final String baseUrl = "http://localhost:8080/address";
 
     @Test
     void a_create() {
         String url = baseUrl + "/create";
-        ResponseEntity<User> postResponse = restTemplate.postForEntity(url, user, User.class);
+        ResponseEntity<Address> postResponse = restTemplate.postForEntity(url, address, Address.class);
         assertNotNull(postResponse);
         assertNotNull(postResponse.getBody());
-        User savedUser = postResponse.getBody();
-        assertEquals(user.getUserId(), savedUser.getUserId());
-        System.out.println("Saved user: " + savedUser);
+        Address savedAddress = postResponse.getBody();
+        assertEquals(address.getId(), savedAddress.getId());
+        System.out.println("Saved address: " + savedAddress);
     }
 
     @Test
     void b_read() {
-        String url = baseUrl + "/read/" + user.getUserId();
+        String url = baseUrl + "/read/" + address.getId();
         System.out.println("URL: " + url);
-        ResponseEntity<User> response = restTemplate.getForEntity(url, User.class);
-        assertEquals(user.getUserId(), response.getBody().getUserId());
+        ResponseEntity<Address> response = restTemplate.getForEntity(url, Address.class);
+        assertEquals(address.getId(), response.getBody().getId());
         System.out.println(response.getBody());
     }
 
     @Test
     void c_update() {
-        User updated = new User.Builder().copy(user).setFirstName("Lihle").build();
+        Address updated = new Address.Builder().copy(address).setCity("Johannesburg").build();
         String url = baseUrl + "/update";
         System.out.println("URL: " + url);
         System.out.println("Post data: " + updated);
-        ResponseEntity<User> response = restTemplate.postForEntity(url, updated, User.class);
+        ResponseEntity<Address> response = restTemplate.postForEntity(url, updated, Address.class);
         assertNotNull(response.getBody());
-        assertEquals("Lihle", response.getBody().getFirstName());
-        System.out.println("Updated user: " + response.getBody());
+        assertEquals("Johannesburg", response.getBody().getCity());
+        System.out.println("Updated address: " + response.getBody());
+    }
+
+    @Test
+    void d_delete() {
+        String url = baseUrl + "/delete/" + address.getId();
+        restTemplate.delete(url);
+
+        ResponseEntity<Address> response = restTemplate.getForEntity(baseUrl + "/read/" + address.getId(), Address.class);
+        assertNull(response.getBody());
+        System.out.println("Deleted address with ID: " + address.getId());
     }
 
     @Test
@@ -77,21 +86,6 @@ public class UserControllerTest {
         System.out.println(response);
         System.out.println(response.getBody());
     }
-
-    @Test
-    void e_delete() {
-        String url = baseUrl + "/delete/" + user.getUserId();
-        System.out.println("URL: " + url);
-
-        restTemplate.delete(url);
-
-        String readUrl = baseUrl + "/read/" + user.getUserId();
-        ResponseEntity<User> response = restTemplate.getForEntity(readUrl, User.class);
-
-        assertNull(response.getBody(), "User should be deleted and not found!");
-        System.out.println("Deleted user with ID: " + user.getUserId());
-    }
 }
-
