+++ src/test/java/za/co/tt/controllerTest/OrderItemControllerTest.java
@@ -1,4 +1,4 @@
-package controllerTest;
+package za.co.tt.controllerTest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
