+++ src/test/java/za/co/tt/factoryTest/LoginRequestFactoryTest.java
@@ -3,7 +3,7 @@
  * Student number: 222582731
  * */
 
-package factoryTest;
+package za.co.tt.factoryTest;
 
 import org.junit.jupiter.api.Test;
 import za.co.tt.domain.LoginRequest;
