+++ src/test/java/za/co/tt/factoryTest/OrderItemFactoryTest.java
@@ -1,4 +1,4 @@
-package factoryTest;
+package za.co.tt.factoryTest;
 
 import org.junit.jupiter.api.Test;
 import za.co.tt.domain.OrderItem;
