+++ src/test/java/za/co/tt/serviceTest/AddressServiceImplTest.java
@@ -2,7 +2,7 @@
  * Author: Yanga Jilaji
  * Student number: 222582731
  * */
-package serviceTest;
+package za.co.tt.serviceTest;
 
 
 import org.junit.jupiter.api.BeforeEach;
