+++ src/test/java/za/co/tt/serviceTest/ReviewServiceTest.java
@@ -1,4 +1,4 @@
-package serviceTest;
+package za.co.tt.serviceTest;
 
 public class ReviewServiceTest {
 }
