+++ src/test/java/za/co/tt/serviceTest/UserServiceImplTest.java
@@ -3,7 +3,7 @@
  * Student number: 222582731
  * */
 
-package serviceTest;
+package za.co.tt.serviceTest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
